+++ bw2data/project.py
@@ -7,13 +7,11 @@ from pathlib import Path
 from threading import ThreadError
 
 import appdirs
-import wrapt
 from bw_processing import safe_filename
 from fasteners import InterProcessLock
 from peewee import BooleanField, DoesNotExist, Model, TextField
 
 from . import config
-from .errors import ReadOnlyProject
 from .filesystem import create_dir
 from .sqlite import PickleField, SubstitutableDatabase
 from .utils import maybe_path
@@ -437,10 +435,3 @@ class ProjectManager(Iterable):
 
 
 projects = ProjectManager()
-
-
-@wrapt.decorator
-def writable_project(wrapped, instance, args, kwargs):
-    if projects.read_only:
-        raise ReadOnlyProject(READ_ONLY_PROJECT)
-    return wrapped(*args, **kwargs)